Extract benchmark report data model
